+++ android/SSPSDK.kt
@@ -1034,4 +1034,4 @@ sealed class SDKError : Exception() {
 }
 
 // SDK version
-const val SDK_VERSION = "0.5.28"
+const val SDK_VERSION = "0.5.29"